Remove unused classes.
