--- de-lib/src/main/java/org/iplantc/de/admin/apps/client/views/grid/AdminAppsGridImpl.ui.xml
@@ -1,18 +1,0 @@
-<ui:UiBinder xmlns:ui='urn:ui:com.google.gwt.uibinder'
-             xmlns:grid="urn:import:com.sencha.gxt.widget.core.client.grid">
-
-    <ui:with field="listStore"
-             type="com.sencha.gxt.data.shared.ListStore" />
-    <ui:with field="cm"
-             type="com.sencha.gxt.widget.core.client.grid.ColumnModel" />
-    <ui:with field="gridView"
-             type="com.sencha.gxt.widget.core.client.grid.GridView">
-        <ui:attributes forceFit="true"
-                       autoFill="true"/>
-    </ui:with>
-
-    <grid:Grid ui:field="grid"
-               store="{listStore}"
-               cm="{cm}"
-               view="{gridView}" />
-</ui:UiBinder>
