+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/apps/Applications.gwt.xml
@@ -37,47 +37,6 @@
         <when-type-is class="org.iplantc.de.apps.client.CommunitiesView.Appearance"/>
     </replace-with>
 
-    <!-- Cells -->
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppCommentCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppCommentCell.AppCommentCellAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppRatingCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppRatingCell.AppRatingCellAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppFavoriteCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppFavoriteCell.AppFavoriteCellAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppNameCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppNameCell.AppNameCellAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppInfoCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppInfoCell.AppInfoCellAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppIntegratorCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppIntegratorCell.AppIntegratorCellAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppStatusCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppStatusCell.AppStatusCellAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppTileCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppTileCell.AppTileCellAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppCardCellDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppCardCell.AppCardCellAppearance"/>
-    </replace-with>
-
-    <replace-with class="org.iplantc.de.theme.base.client.apps.cells.AppDotMenuDefaultAppearance">
-        <when-type-is class="org.iplantc.de.apps.client.views.list.cells.AppDotMenuCell.AppDotMenuAppearance"/>
-    </replace-with>
-    
     <!-- Submit -->
     <replace-with class="org.iplantc.de.theme.base.client.apps.submit.SubmitAppViewDefaultAppearance">
         <when-type-is class="org.iplantc.de.apps.client.SubmitAppForPublicUseView.SubmitAppAppearance"/>
