--- de-lib/src/main/java/org/iplantc/de/theme/base/client/apps/cells/AppDotMenu.gss
@@ -1,3 +1,0 @@
-.dot_menu {
-	cursor: pointer;
-}
