--- de-lib/src/main/java/org/iplantc/de/theme/base/client/apps/cells/AppInfoCell.gss
@@ -1,3 +1,0 @@
-.app_info {
-	cursor: pointer;
-}
