--- de-lib/src/main/java/org/iplantc/de/theme/base/client/apps/cells/AppNameCell.gss
@@ -1,35 +1,0 @@
-.appHyperlinkName {
-    cursor: pointer;
-    text-decoration: none;
-    color: black;
-    font-size: 12px;
-    height: 15px;
-    padding-top: 2px;
-    padding-bottom: 2px;
-}
-
-.appHyperlinkName:hover {
-    text-decoration: underline;
-}
-
-.appName {
-    text-decoration: none;
-    color: black;
-    font-size: 12px;
-    height: 15px;
-    padding-top: 2px;
-    padding-bottom: 2px;
-}
-
-.appDisabled {
-    cursor: hand;
-    font-size: 12px;
-    height: 15px;
-    padding-top: 2px;
-    padding-bottom: 2px;
-}
-
-.appDisabled:hover {
-    cursor: hand;
-}
-
