--- de-lib/src/main/java/org/iplantc/de/theme/base/client/apps/cells/FavoriteCellStyle.gss
@@ -1,35 +1,0 @@
-.favorite {
-    gwt-sprite: "favIcon";
-    cursor: pointer;
-    float: left;
-}
-
-.favorite:hover {
-    gwt-sprite: "favIconDelete";
-    cursor: pointer;
-    float: left;
-}
-
-.favoriteDisabled {
-    gwt-sprite: "disabledFavIcon";
-    cursor: default;
-    float: left;
-}
-
-.favoriteDisabled:hover {
-    gwt-sprite: "disabledFavIcon";
-    cursor: default;
-    float: left;
-}
-
-.favoriteAdd {
-    gwt-sprite: "disabledFavIcon";
-	cursor: pointer;
-    float: left;
-}
-
-.favoriteAdd:hover {
-    gwt-sprite: "favIconAdd";
-    cursor: pointer;
-    float: left;
-}
