+++ de-lib/src/main/java/org/iplantc/de/theme/base/client/diskResource/grid/cells/DotMenu.gss
@@ -1,0 +1,3 @@
+.dot_menu {
+    cursor: pointer;
+}
